Make a private function public
It is better to do this for functions that are added to hooks, because
those are anyhow used for "public" purposes: they are not limited to
some internal operation of the package.

The practical advantage for the user is that we make a promise to keep
this function working and to document any changes to its behaviour.