Do not disable menu bar beframe-buffer-menu entry even with single frame
The reason is that even if we only have a single frame, it can still
have its own beframed buffer list.  This happens when we create a new
frame, close the others, and now have access only to the buffers
visited from the new frame.  As always, the global buffer list is
still available, though it has to be accessed with commands other than
those of Beframe (e.g. 'consult-buffer' from the 'consult' package).